refactor : container-view structure 구조로 변경
1. container : state 변경 등 비지니스 로직이 포함된 component / view가 container에 포함됨
2. view : props를 입력받아 화면에 표시되는 component
