+++ src/containers/OnBoardingContainerStyle.css
@@ -1,0 +1,22 @@
+.double-logo {
+  animation-iteration-count: 1;
+  animation-duration: 3s;
+  animation-name: double;
+  animation-fill-mode: forwards;
+}
+header {
+  width: 100%;
+  height: 5px;
+}
+
+@keyframes double {
+  0% {
+    scale: 1;
+  }
+  50% {
+    scale: 2;
+  }
+  100% {
+    scale: 1;
+  }
+}
