+++ src/pages/Onbaording/OnBoarding.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import OnBoardingContainer from '../../containers/OnboardingContainer';
+
+function OnBoarding(): JSX.Element {
+  return <OnBoardingContainer />;
+}
+
+export default OnBoarding;
