--- src/pages/onboarding/OnBoarding.css
@@ -1,49 +1,0 @@
-.absolute-div {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  background-color: white;
-  z-index: 9999;
-  animation-iteration-count: 1;
-  animation-duration: 3s;
-  animation-name: fade-out;
-  animation-fill-mode: forwards;
-}
-.double-logo {
-  animation-iteration-count: 1;
-  animation-duration: 3s;
-  animation-name: double;
-  animation-fill-mode: forwards;
-}
-header {
-  width: 100%;
-  height: 5px;
-}
-
-@keyframes fade-out {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-    display: none;
-  }
-}
-
-@keyframes double {
-  0% {
-    scale: 1;
-  }
-  50% {
-    scale: 2;
-  }
-  100% {
-    scale: 1;
-  }
-}
